refactor: CheckInDateTime을 CheckInDateTime와 CheckInTime으로 분리
