+++ src/main/java/domain/CheckInDateTime.java
@@ -1,35 +1,19 @@
 package domain;
 
-import exception.AppException;
+public class CheckInDateTime {
+    private final CheckInDate checkInDate;
+    private final CheckInTime checkInTime;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.Objects;
-
-public class CheckInDateTime implements Comparable<CheckInDateTime> {
-    private final LocalDate checkInDate;
-    private final LocalTime checkInTime;
-
-    public static final LocalTime CAMPUS_START_TIME = LocalTime.of(8, 0);
-    public static final LocalTime CAMPUS_END_TIME = LocalTime.of(23, 0);
-
-    private CheckInDateTime(LocalDate checkInDate, LocalTime checkInTime) {
-        validateCampusTime(checkInTime);
+    private CheckInDateTime(CheckInDate checkInDate, CheckInTime checkInTime) {
         this.checkInDate = checkInDate;
         this.checkInTime = checkInTime;
     }
 
-    public static CheckInDateTime of(LocalDate checkInDate, LocalTime checkInTime) {
+    public static CheckInDateTime of(CheckInDate checkInDate, CheckInTime checkInTime) {
         return new CheckInDateTime(checkInDate, checkInTime);
     }
 
-    private void validateCampusTime(LocalTime checkInTime) {
-        if (checkInTime.isBefore(CAMPUS_START_TIME) || checkInTime.isAfter(CAMPUS_END_TIME)) {
-            throw new AppException("캠퍼스 이용 시간은 08시부터 23시까지 입니다.");
-        }
-    }
-
-    @Override
+    /*    @Override
     public int compareTo(CheckInDateTime o) {
         return this.checkInDate.compareTo(o.checkInDate);
     }
@@ -45,5 +29,5 @@ public class CheckInDateTime implements Comparable<CheckInDateTime> {
     @Override
     public int hashCode() {
         return Objects.hash(checkInDate);
-    }
+    }*/
 }
